pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "feature-attainment"
version = "0.1.0"
description = "Fork of purejaxrl for feature attainment exps"
readme = "README.md"
authors = [
    {name = "Oliver Diamond"},
]
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "jax",
    "PyExpUtils-andnp~=7.0",
    "jaxlib",
    "flax",
    "optax",
    "distrax",
    "flashbax",
    "gymnax",
    "brax",
    "navix",
    "numpy",
    "wandb",
    "filelock",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
test = [
    "pytest",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "README.md",
    "pyproject.toml",
]